Support dual mode cache
